requirements.txt
Django==2.0.1
pyserial==3.4
python-decouple==3.1
pytz==2017.3
virtualenv==15.1.0